Fix tests
Roles now comes in a dict instead of a list.
proc_name was changed to aio_etcd, but this is the etcd binary, should
not be changed.
Port 6001 is used by X
Setting a role by just setting the name does not work, remove test.
It is not needed to get the full etcd path to call it.